KASM-4323 KASM-4320 Layering and Locales

See merge request kasm-technologies/internal/workspaces-images!111

+++ src/ubuntu/install/minetest/install_minetest.sh
@@ -21,8 +21,9 @@ EOL
 chmod +x /usr/bin/minetest
 
 # Cleanup
-apt-get autoclean
-rm -rf \
+if [ -z ${SKIP_CLEAN+x} ]; then
+  apt-get autoclean
+  rm -rf \
     /var/lib/apt/lists/* \
-    /var/tmp/* \
-    /tmp/*
+    /var/tmp/*
+fi
