+++ src/ubuntu/install/vs_code/install_vs_code.sh
@@ -20,8 +20,9 @@ apt-get install -y python3-setuptools \
                    python3-virtualenv
 
 # Cleanup
-apt-get autoclean
-rm -rf \
+if [ -z ${SKIP_CLEAN+x} ]; then
+  apt-get autoclean
+  rm -rf \
     /var/lib/apt/lists/* \
-    /var/tmp/* \
-    /tmp/*
+    /var/tmp/*
+fi
